Site doc generator: respect Javadoc on type for smallrye-config
Considers the Javadoc of the "most relevant" config-mapping type and adds it as markdown to the generated docs.
