+++ servers/quarkus-common/src/main/java/org/projectnessie/quarkus/config/QuarkusDynamoDBConfig.java
@@ -19,6 +19,11 @@ import io.quarkus.runtime.annotations.StaticInitSafe;
 import io.smallrye.config.ConfigMapping;
 import java.util.Optional;
 
+/**
+ * When setting {@code nessie.version.store.type=DYNAMODB} which enables DynamoDB as the version
+ * store used by the Nessie server, the following configurations are applicable in combination with
+ * {@code nessie.version.store.type}.
+ */
 @StaticInitSafe
 @ConfigMapping(prefix = "nessie.version.store.persist.dynamodb")
 public interface QuarkusDynamoDBConfig {
