+++ servers/quarkus-common/src/main/java/org/projectnessie/quarkus/config/QuarkusRocksConfig.java
@@ -22,6 +22,11 @@ import io.smallrye.config.WithName;
 import java.nio.file.Path;
 import org.projectnessie.versioned.storage.rocksdb.RocksDBBackendBaseConfig;
 
+/**
+ * When setting {@code nessie.version.store.type=ROCKSDB} which enables RocksDB as the version store
+ * used by the Nessie server, the following configurations are applicable in combination with {@code
+ * nessie.version.store.type}.
+ */
 @StaticInitSafe
 @ConfigMapping(prefix = "nessie.version.store.persist.rocks")
 public interface QuarkusRocksConfig extends RocksDBBackendBaseConfig {
